refactor(const): Move color_to_hex() from render.py to const.py
