+++ custom_components/eink_dashboard/widgets/heading.py
@@ -15,6 +15,7 @@ from ..const import (
     PADDING,
     DisplayConfig,
     Widget,
+    color_to_hex,
 )
 from ..svg_render import _mdi_svg_filter
 from ._helpers import (
@@ -90,7 +91,6 @@ def _build_heading_context(
         _compute_metrics,
         _device_class_icon,
         _load_font,
-        color_to_hex,
     )
 
     x = widget.get("x", PADDING)
